+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 ## Features
 
+* Remove deprecated classes and methods
 * Add techdivision/import-product#146
 * Add techdivision/import-cli-simple#216
 * Remove unnecessary identifiers from configuration